Remove gitlab instance for cli tests
